Expanding validation checks to catch additional issues before launching into the transfer loops, identify whether existing rsync processes are running and skipping the post-transfer comparison if 'comm' isn't available on the system.
Added the option to determine the number of processors that should be used on a system (with appropriate validation checks).

+++ rsync_data_transfer.sh
@@ -12,15 +12,12 @@
 #-----------------#
 #    VARIABLES    #
 #-----------------#
-
-USER=""																			## The username required to establish a secure shell (ssh) connection to the remote host:
-#read -p "Username: " USER														## User prompts for interactive invocation. Commented out by default.
-
+USER=""																			## The username required to establish a secure shell (ssh) connection to the remote host
+#read -p "Username: " USER														## User prompts for interactive invocation. Commented out by default
 REMOTE_HOST=""																	## The remote host name or IP address
-#read -p "Remote host: " REMOTE_HOST											## User prompts for interactive invocation. Commented out by default.
-
-REMOTE_DIR=""																	## The directory on the remote server that the data will be transferred into:
-#read -p "Remote directory: " REMOTE_DIR										## User prompts for interactive invocation. Commented out by default.
+#read -p "Remote host: " REMOTE_HOST											## User prompts for interactive invocation. Commented out by default
+REMOTE_DIR=""																	## The directory on the remote server that the data will be transferred into
+#read -p "Remote directory: " REMOTE_DIR										## User prompts for interactive invocation. Commented out by default
 
 
 	#-------------------------------------------------------------------------------#
@@ -35,6 +32,7 @@ REMOTE_DIR=""																	## The directory on the remote server that the dat
 #-----------------#
 
 validation_checks() {
+## Checking that the variables in this script (above) have been populated:
 if [[ -z ${USER} ]] || [[ -z ${REMOTE_HOST} ]] || [[ -z ${REMOTE_DIR} ]]
 then
 	echo -e "\nERROR:\tThe following variables have not been defined within the script:\n\tUser:\t\t${USER}\n\tRemote Host:\t${REMOTE_HOST}\n\tRemote Directory:\t${REMOTE_DIR}"
@@ -49,6 +47,13 @@ then
     TERMINATE="true"
 fi
 
+## Validating the integer provided for the number of processors: 
+if [[ ${PROCS} -gt $(nproc) ]]
+then
+	echo -e "\nERROR:\tThe number of processors specified is greater than the number of CPU cores on the local server.\n"
+	TERMINATE="true"
+fi
+
 ## Checking for the existence of a trailing slash on the provided directory path:
 DIR_SLASH=$(echo ${SOURCE_DIR: -1})
 ## If there is a trailing slash, let's remove it from the path as the rsync syntax below includes the slash (and we don't want duplicate slashes):
@@ -58,7 +63,7 @@ then
 fi
 
 ## Validation that passwordless authentication is enabled between source and destination servers (e.g. using ssh keys):
-ssh -o PasswordAuthentication=no -o BatchMode=yes ${USER}@${REMOTE_HOST} exit 2>&1 /dev/null
+ssh -o PasswordAuthentication=no -o BatchMode=yes ${USER}@${REMOTE_HOST} exit > /dev/null
 ## An unsuccessful attempt will return a non-zero error code, which will fail the following check:
 if [[ $? == 0 ]]
 then 
@@ -77,8 +82,15 @@ else
 	TERMINATE="true"
 fi
 
+## Looking for pre-existing rsync processes on the local server:
+if [[ $(ps -e -o cmd | awk '$1=="rsync"') ]]
+then
+	echo -e "\nADVISORY:\tThere are running rsync processes on the local server:"
+	ps -e -o psr,cmd,pid | awk '$2=="rsync"'
+fi
+
 ## Checking rsync is installed on the remote server:
-ssh ${USER}@${REMOTE_HOST} 'command -v rsync' 2>&1 /dev/null
+ssh ${USER}@${REMOTE_HOST} 'command -v rsync' > /dev/null
 ## An unsuccessful attempt will return a non-zero error code, which will fail the following check:
 if [[ $? == 0 ]]
 then
@@ -95,6 +107,13 @@ then
 	TERMINATE="true"
 fi
 
+## Validating that the variable containing the number of processors is populated correctly. If 'nproc' isn't available, the variable value will be -1 and this will cause problems...
+if [[ ${NUM_CPUS} == "-1" ]]
+then
+	echo "Unable to accurately determine number of processors using 'nproc'. Make this program available (and in '$PATH') or manually amend the NUM_CPUS variable to proceed.\n"
+	TERMINATE="true"
+fi
+
 ## If any of the prior validation checks fail, then the help() function will be called and the script will exit: 
 if [[ ${TERMINATE} == "true" ]]
 then
@@ -103,9 +122,10 @@ then
 fi
 }
 
+## Defining the help function to be invoked if no arguments provided at runtime, or the validation checks fail:
 help() {
-	echo -e "\nHELP STATEMENT\nPlease execute the script specifying the parameters for source directory '-d' and number of parallel threads '-t' as an integer (i.e. not a floating point number)."
-	echo -e "\nExample usage:\n\n\t $ /path/to/script.sh -d /directory/with/files/to/send -t 32\n"
+	echo -e "\nHELP STATEMENT\nPlease execute the script specifying the parameters for source directory '-d', the number of processors '-p' as either 'all' or an integer, and the number of parallel threads '-t', also as an integer (i.e. not a floating point number)."
+	echo -e "\nExample usage:\v\t $ /path/to/script.sh -d /directory/with/files/to/send -p 4 -t 32\n"
 	echo -e "\nPackages & commands required:\tssh; nproc; ps; awk; sed; rsync (on local server); rsync (on remote server); taskset; comm\n"
 }
 
@@ -116,9 +136,7 @@ help() {
 
 TIMER_START=$(date +%s)															## Capturing the starting second count to be used to calculate the wall time
 
-echo -e "\nBeginning validation...\n"
-
-while getopts "hd:t:" OPTION
+while getopts "hd:t:p:" OPTION
 do
 case "$OPTION"
 in
@@ -141,35 +159,54 @@ in
 			help
 			exit 1
 		fi 
-	;;									
+	;;
+	p) PROCS=${OPTARG}
+		if [[ ${PROCS} =~ ALL|All|all ]]										## Determining whether user input determines that all server CPUs will be used for data transfer 
+		then
+			echo -e "VALIDATED:\tAll system processors selected."
+			PROCS=$(nproc)
+		elif ((${PROCS})) 2> /dev/null 											## Checking that the thread value provided is an integer (not a string nor a float)
+			then
+				echo -e "VALIDATED:\tProcessor value provided is an integer."
+			else
+				echo -e "\nERROR:\tPlease specify an integer (whole number) for the number of processors to be used for the data transfer, or 'all' to specify all processors.\n"
+				help
+				exit 1	
+		fi
     h | *) help && exit 1														## Capturing all other input; providing the help() statement for non-ratified inputs
 	;;
 esac
 done
 
-validation_checks																## Calling the validation_checks function
-
 ## Creating the runtime variables:
 TOTAL_TASKS=$(find ${SOURCE_DIR} -type f | wc -l)								## The total number of files in the supplied directory path to be transferred
 FILE_QUEUE=( $(ls ${SOURCE_DIR}) )												## Creating a variable array that contains the file names that are to be transferred
-NUM_CPUS=$(( $(nproc) - 1 ))													## The number of CPUs to be used for transfers on the source server, less 1 as we number from 0
+NUM_CPUS=$(( ${PROCS} - 1 ))													## The number of CPUs to be used for transfers on the source server, less 1 as we number from 0
 FILE_INDEX="0"																	## A simple file counter used to measure the number of tasks being undertaken
 
-echo -e "\nSource directory:\t\t${SOURCE_DIR}\nRemote directory:\t\t${REMOTE_DIR}\nNumber of tasks:\t\t${TOTAL_TASKS}\nNumber of processors:\t\t${NUM_CPUS}\nThread count:\t\t\t${THREADING}\n"
+echo -e "\nBeginning validation...\n"
+validation_checks																## Calling the validation_checks function
+
+echo -e "
+Source directory:\t\t${SOURCE_DIR}
+Remote directory:\t\t${REMOTE_DIR}
+Number of tasks:\t\t${TOTAL_TASKS}
+Number of processors:\t\t${PROCS}
+Thread count per CPU:\t\t${THREADING}\n"										## Printing the defined variables to stdout to create a record of the conditions
 
 ## Sending table headings to stdout for transfer information:
 echo -e "\nHOSTNAME\t\t\t\tCPU\t\tTASK\t\tTHREAD\t\tFILE"
 
 while true
 do
-	## Cycling the available CPUs in the source system:
+	## Cycling the available CPUs on the local server:
     for CPU in $(seq 0 ${NUM_CPUS})
     do
 		## Tracking that we still have outstanding tasks to complete:
         if [ ${FILE_INDEX} -lt ${TOTAL_TASKS} ]
         then
 			## Running a check to see whether any rsync processes are running on the specific processor:
-            CHECK="$(ps -e -o psr,cmd | awk -v aCPU=${CPU} '$1==aCPU' | awk '$2=="rsync"')"
+            CHECK=$(ps -e -o psr,cmd | awk -v aCPU=${CPU} '$1==aCPU' | awk '$2=="rsync"')
 
 			## If the variable is empty (and thus no process running), bind an rsync operation to the specific processor for the next file in the FILE_QUEUE: 
             if [[ -z ${CHECK} ]]
@@ -206,40 +243,45 @@ do
 				echo -n "Remaining processes: `pidof rsync | wc -w`"
   				echo -n -e "\e[0K\r"
 			done
+			echo -e "All processes complete."
 
 			## Checking for differences between source target directories:
-			echo -e "\n\n\nChecking for the differences between source & remote directories..."
-			FILE_LISTS="/dev/shm/data-transfer-file-list"
-			ls ${SOURCE_DIR} | sort > ${FILE_LISTS}.source																	## Capturing the contents of the source directory and storing in a temp file on local memory
-			ssh ${USER}@${REMOTE_HOST} "ls ${REMOTE_DIR} | sort" > ${FILE_LISTS}.remote										## Capturing the contents of the remote directory and storing in a temp file on local memory
-			DIR_COMPARISON=( $(comm -23 ${FILE_LISTS}.source ${FILE_LISTS}.remote) )										## Comparing the source & remote directories from the temp files just created, and storing any differences in a variable array
-			
-			if [[ -n ${DIR_COMPARISON} ]]																					## A query on the variable with '-n' sees whether there is a value set. If there is, follow the loop... 
+			if [[ -x $(command -v comm) ]]
 			then
-				if [[ $(ls ${SOURCE_DIR} | wc -l) == ${TOTAL_TASKS} ]]														## Checking to see whether the current number of files in the source directory matches $TOTAL_TASKS, generated earlier in the script
-				then
-					echo -e "\nNot all files have been transferred during this operation."
-				else
-					echo -e "\nThere is a difference in the number of files present than when the transfer was initiated."
-				fi
-				echo -e "\nThe following files exist on the source but not on the destination:"
-				for DIFF_FILE in ${DIR_COMPARISON[*]}																		## Looping through the variable array and printing the contents to stdout
-				do 
-					echo -e "\t${DIFF_FILE}"
-				done	
-				echo -e "\nYou can re-run the script and rsync will send only those files that do not exist on the remote directory."
+				echo -e "\v\vChecking for the differences between source & remote directories..."
+				FILE_LISTS="/dev/shm/data-transfer-file-list"																	## Storing the file lists in memory on the local server (should be pretty small)
+				ls ${SOURCE_DIR} | sort > ${FILE_LISTS}.source																	## Capturing the contents of the source directory and storing in a temp file on local memory
+				ssh ${USER}@${REMOTE_HOST} "ls ${REMOTE_DIR} | sort" > ${FILE_LISTS}.remote										## Capturing the contents of the remote directory and storing in a temp file on local memory
+				DIR_COMPARISON=( $(comm -23 ${FILE_LISTS}.source ${FILE_LISTS}.remote) )										## Comparing the source & remote directories from the temp files just created, and storing any differences in a variable array
 			
-			else																											## The alternative, assuming there is no value stored in $DIR_COMPARISON 
-				echo -e "\nThe source and remote directories are in sync - all files were successfully transferred."
+				if [[ -n ${DIR_COMPARISON} ]]																					## A query on the variable with '-n' sees whether there is a value set. If there is, follow the loop... 
+				then
+					if [[ $(ls ${SOURCE_DIR} | wc -l) == ${TOTAL_TASKS} ]]														## Checking to see whether the current number of files in the source directory matches $TOTAL_TASKS, generated earlier in the script
+					then
+						echo -e "\nNot all files have been transferred during this operation."
+					else
+						echo -e "\nThere is a difference in the number of files present than when the transfer was initiated."
+					fi
+					echo -e "\nThe following files exist on the source but not on the destination:"
+					for DIFF_FILE in ${DIR_COMPARISON[*]}																		## Looping through the variable array and printing the contents to stdout
+					do 
+						echo -e "\t${DIFF_FILE}"
+					done	
+					echo -e "\nYou can re-run the script and rsync will send only those files that do not exist on the remote directory."
+				
+				else																											## The alternative, assuming there is no value stored in $DIR_COMPARISON 
+					echo -e "\nThe source and remote directories are in sync - all files were successfully transferred."
+				fi
+				rm ${FILE_LISTS}.source ${FILE_LISTS}.remote																	## Being good citizens and tidying up after ourselves
+			else
+				echo -e "The 'comm' comparison program is not available - skipping post-transfer directory comparison...\n"
 			fi
-			rm ${FILE_LISTS}.source ${FILE_LISTS}.remote																	## Being good citizens and tidying up after ourselves
+			echo -e "\vOPERATION COMPLETE: Submitted ${FILE_INDEX} files for transfer to ${REMOTE_HOST}:${REMOTE_DIR}\v"
 
-			echo -e "\n\nOPERATION COMPLETE: Submitted ${FILE_INDEX} files for transfer to ${REMOTE_HOST}:${REMOTE_DIR}\n\n"
-
-			TIMER_END=$(date +%s)																							## Capturing the end second count
-			TIMER_DIFF_SECONDS=$(( ${TIMER_END} - ${TIMER_START} ))															## Calculating the difference
-			TIMER_READABLE=$(date +%H:%M:%S -ud @${TIMER_DIFF_SECONDS})														## Converting the second delta into a human readable time format (HH:MM:SS)
-			echo -e "Date:\t\t`date "+%a %d %b %Y"`\nWall time:\t\t${TIMER_READABLE}\n"										## And printing it to stdout with the date
+			TIMER_END=$(date +%s)																								## Capturing the end second count
+			TIMER_DIFF_SECONDS=$(( ${TIMER_END} - ${TIMER_START} ))																## Calculating the difference
+			TIMER_READABLE=$(date +%H:%M:%S -ud @${TIMER_DIFF_SECONDS})															## Converting the second delta into a human readable time format (HH:MM:SS)
+			echo -e "Date:\t\t`date "+%a %d %b %Y"`\nWall time:\t\t${TIMER_READABLE}\n"											## And printing it to stdout with the date
 
 			exit 0
         fi
